refactor: move ApprovalState to common/hooks
