+++ apps/cowswap-frontend/src/legacy/state/types.ts
@@ -3,7 +3,6 @@ export enum ConfirmOperationType {
   UNWRAP_WETH,
   APPROVE_TOKEN,
   REVOKE_APPROVE_TOKEN,
-  ORDER_SIGN,
   ORDER_CANCEL,
   CONVERT_VCOW,
   CLAIM_VESTED_COW,
